refactor: ReservationCreateRequest를 controller 패키지로 이동
