+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -5,7 +5,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.ReservationCreateRequest;
+import roomescape.controller.ReservationCreateRequest;
 import roomescape.entity.Reservation;
 
 import java.sql.PreparedStatement;
